Fixed Icon Customization
.exe Icons will take the icon of the .exe application until a user changes it in the context menu under "Properties".

Folder/Directory Icons will take the icon of a system folder until a user changes it in the context menu under "Properties".

Added a visual UI to view all available icons that users can double click and by doing that, set the selected icon as a replacement for the .exe or folder the user wants to manipulate.