Tweaked NitroDockX...
Added: user control to resize the docked icons and adjust the spacing between them in the Configuration Panel.

Added: the ability to move the icons and reorder them using the middle mouse wheel. Click the icon with the middle mouse wheel button and use the wheel to move it.

Added: users can now change the icons of the apps and directories moved onto the dock. Icons can be in .png and .ico formats and this allows for a much more customized experience.

***On that note: I am experimenting with skinning so this may take a while. As of late the dock works however there is no persistence; meaning the dock, if closed will lose all your setup for now. I plan on implementing an ini configuration file that saves buttons and the shortcutc so that your dock functions on restart if the application closes down.***