Added Some Visuals for Devs
Added: Icon Containers that prevent the icons from expanding outside them. This will aid in skinning the Dock

Added: the ability to add background color, textures (images), and transparency to the Icon Containers so users can skin the Dock in different ways which can be as complicated or as easy as they want it to be

Fixed: an issue with moving icons up and down and side to side on the Dock. This was an OCD fix, previously you could play with the icons and get it to work, but now its more definitive

Moved: context menu items around so they make sense and only appear on the Dock Elements they were assigned to. So people don't get confused.

Existing Bugs: at max zoom level all icons blur and it may cause an issue with a possible crash. At min zoom level all icons minimize just fine except the configure user settings button.

Intended: there are no labels on the dock, this was done for aesthetics. If you're a user and you place items on the dock, then you should know what items you docked. Labels just look messy and I never liked them.