Fixed A Major Headache
So for those of you who don't know, I've been struggling for a few days to fix a problem where the dock would go off-screen when changes were applied.

This issue only affected those with multiple monitors and since I have four, this was a real migrane to fix. Anyway without further tension... I hope some of you get use out of this.

It's open source so if you have issues, contact  me or adapt the code to your liking and send me screenshots!